Revert "fix: rename (#4)"
This reverts commit 6e32fe799f151c941de7119dcb7eaf30ee9cced2.
