+++ src/application/modules/users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '@/infrastructure/database/entities/user.entity';
+import { User } from '@/infrastructure/db/entities/user.entity';
 import { UsersResolver } from '../graphql/user/resolvers/users.resolver';
 import { UsersService } from '@/domain/user/services/users.service';
-import { UserRepository } from '@/infrastructure/database/repositories/user.repository';
+import { UserRepository } from '@/infrastructure/db/repositories/user.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
