+++ src/infrastructure/db/shared/database/entities/base.entity.ts
