+++ src/infrastructure/db/shared/database/entities/deletable.entity.ts
